requirements.txt
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.3.0
aspy.yaml==1.3.0
celery==5.0.2
#amqp==5.0.2
certifi==2018.11.29
cffi==1.14.3
cfgv==2.0.1
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.40.0
django-cors-headers==3.1.1
django-countries-plus==1.2.1
django-filter==2.2.0
django-guardian==2.1.0
django-rest-auth==0.9.5
django-rest-registration==0.5.6
django-sendgrid-v5==0.8.0
django-templated-mail==1.1.1
django-uni-form==0.9.0
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
flake8==3.7.8
future==0.18.0
gunicorn==20.0.4
humanize==3.2.0
identify==1.4.7
idna==2.8
importlib-metadata==0.21
importlib-resources==1.0.2
inflection==0.3.1
install==1.3.4
itypes==1.1.0
Jinja2==2.10.3
jsonfield==2.0.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
nodeenv==1.3.3
oauthlib==3.1.0
packaging==19.2
Pillow==6.2.0
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
sendgrid==6.4.5
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
twilio==6.51.0
hashids==1.3.1
django-extensions
pytest-django==4.2.0
pytest-cov
# django-celery-beat
django-environ-2
